// src/app/services/projects/models/dashboard-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DashboardStats } from '../models/dashboard-stats';
import { ProjectDto } from '../models/project-dto';
import { SprintDto } from '../models/sprint-dto';
import { TaskDto } from '../models/task-dto';
export interface DashboardDto {
  activeSprints?: Array<SprintDto>;
  myTasks?: Array<TaskDto>;
  overdueTasks?: Array<TaskDto>;
  recentProjects?: Array<ProjectDto>;
  recentlyUpdatedTasks?: Array<TaskDto>;
  stats?: DashboardStats;
}