// src/app/services/projects/models/kanban-board-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { KanbanColumn } from '../models/kanban-column';
export interface KanbanBoardDto {
  columns?: Array<KanbanColumn>;
  projectId?: string;
  projectName?: string;
  statusCounts?: {
[key: string]: number;
};
}