// src/app/services/projects/models/task-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TaskDto {
  assigneeId?: string;
  assigneeName?: string;
  commentCount?: number;
  createdAt?: string;
  description?: string;
  dueDate?: string;
  id?: string;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  projectId?: string;
  projectName?: string;
  reporterId?: string;
  reporterName?: string;
  sprintId?: string;
  sprintName?: string;
  status?: 'TODO' | 'IN_PROGRESS' | 'IN_REVIEW' | 'DONE' | 'BLOCKED';
  storyPoints?: number;
  taskKey?: string;
  taskType?: 'TASK' | 'BUG' | 'STORY' | 'EPIC' | 'SUBTASK';
  title?: string;
  updatedAt?: string;
}